+++ src/components/designSystem/Filters/utils.ts
@@ -34,6 +34,7 @@ import { TranslateFunc } from '~/hooks/core/useInternationalization'
 
 import {
   ActivityLogsAvailableFilters,
+  AMOUNT_INTERVALS_TRANSLATION_MAP,
   AmountFilterInterval,
   AnalyticsInvoicesAvailableFilters,
   AvailableFiltersEnum,
@@ -390,13 +391,6 @@ export const formatFiltersForActivityLogsQuery = (searchParams: URLSearchParams)
   }
 
   return formatted
-}
-
-export const AMOUNT_INTERVALS_TRANSLATION_MAP = {
-  [AmountFilterInterval.isBetween]: 'text_1734774653389kvylgxjiltu',
-  [AmountFilterInterval.isEqualTo]: 'text_1734774653389pt3rhh3lspa',
-  [AmountFilterInterval.isUpTo]: 'text_1734792781750cot2uyp6f1x',
-  [AmountFilterInterval.isAtLeast]: 'text_17347927817503hromltntvm',
 }
 
 export const formatActiveFilterValueDisplay = (
